Move document addition functionality to separate file
This clears up module.c a bit more
